+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Yandex Turbo Pages Change Log
 =============================
 
-Dev-master
+1.1.2 November 15, 2018
 --------------------
 * Enh: added accordion method into Content helper
 * Enh: added rating method into Content helper